[ADD] base classes
